Submission for Predict Bikes
I faced some difficulties when trying to convert my .ipynb file to an HTML file. The conversion process was not going smoothly, and I encountered errors or unexpected results. In order to resolve this issue, I decided to take an alternative approach.
Instead of directly converting the .ipynb file to an HTML file, I created a .zip file that contains the .md (Markdown) file. Markdown is a lightweight markup language commonly used for formatting text. By using the .md file within the .zip archive, I can still share the contents of my .ipynb file, albeit in a different format.